Removed redundant components. Modified the handling of App ID.
Modified Advertiser callback's AsyncApiCaller to make use of Publisher's UrlBuilder.
Deleted AdvertiserHostInfo, which contained references to program_id.
Renamed PublisherHostInfo to HostInfo.
Unified HostInfo with DeviceInfo.
Modified the handling of App ID to let it be a String instead of an always numeric value.
Throws a RuntimeException if App ID is not set.