Refactored. Modified the way a successful response from the server to the advertiser callback affects subsequent invocations of the advertiser callback method.
Moved version information to a the new SponsorPay class.
Renamed AsyncAPICaller to AdvertiserCallbackSender.
The Advertiser Callback is now always sent regardless of whether a successful response from the server had been already received. But a new parameter in the call with the key "answer_received" is set to 1 if the callback had already been successful or to 0 otherwise.