* OfferWallActivity is set as the owner of the error dialog whose .show() method was triggering the exception, If still an exception is thrown, we catch it silently. This is because this exception should mean that the activity is not running anymore, and in this case the error dialog shouldn't be shown.
* catch the OutOfMemoryError thrown when the SDK is trying to send a request to the Virtual Currency Server and return a generic error to your listener

For the BadTokenException thrown from the OfferWallActivity, it was being thrown when trying to show an error dialog. The changes are the following:
I make sure the OfferWallActivity is set as the owner of the error dialog whose .show() method was triggering the exception,
If still an exception is thrown, we catch it silently. This is because this exception should mean that the activity is not running anymore, and in this case the error dialog shouldn't be shown.
For the OutOfMemoryError thrown from the SDK, although this error is thrown from the SDK it doesn't necessarily mean that it's caused by the SDK itself. The SDK just tries to allocate memory to do something and there's not enough memory available on the heap. That memory can have been consumed by something else. If you do suspect the SDK is consuming too much memory in your application please give me more details and we'll find a solution. The quick and dirty fix I've implemented for this is to catch the OutOfMemoryError thrown when the SDK is trying to send a request to the Virtual Currency Server and return a generic error to your listener callback. The fix I've implemented should prevent this error thrown at this particular point from crashing your app, but won't solve the underlying issue. Again, if you suspect the SDK is causing your app from running out of memory let's speak about it.