Fixed bug: Add check for null cookies on the Interstitial request.
In the rare event the request for the Interstitial returns a response which indicates that the Interstitial is available but contains no cookie headers, treat it like an Interstitial Not Available case and don't show the Interstitial Activity.

If the Interstitial Activity is launched but no cookies have been passed to it, silently fail and close the Interstitial instead of letting an exception be thrown.